refactor: extract composer (input box) into ui/composer.rs (#357)
Third slice of the ui.rs decomposition. Moved draw_input (~135 lines)
into its own submodule. The composer reads input/reply/edit/attachment
state, renders the bordered text pane with title and badge, and
positions the terminal cursor in Insert mode.

ui/mod.rs: 4984 -> 4849 lines.
